refactor(coding-agent): extract agent-session queue module
Move the queued-message classification + restoration helpers (isAdvisorCard,
isUserQueuedMessage, isDisplayableQueuedMessage, isHiddenUserCompanion,
queueChipText, toRestoredQueuedMessage, the internal queuedText/ImageContent and
magic-keyword notice set, plus IMAGE_ATTACHMENT_DESCRIPTION_TYPE) into sibling
agent-session-queue.ts. Pure functions over AgentMessage; AgentSession imports
back the eight symbols it uses. The public RestoredQueuedMessage type is
re-exported from the old path, so the runtime export surface is byte-identical.

Op: compress
