+++ packages/coding-agent/src/session/agent-session.ts
@@ -319,6 +319,16 @@ import {
 	PERMISSION_OPTIONS_BY_ID,
 	PERMISSION_REQUIRED_TOOLS,
 } from "./agent-session-permissions";
+import {
+	IMAGE_ATTACHMENT_DESCRIPTION_TYPE,
+	isAdvisorCard,
+	isDisplayableQueuedMessage,
+	isHiddenUserCompanion,
+	isUserQueuedMessage,
+	queueChipText,
+	type RestoredQueuedMessage,
+	toRestoredQueuedMessage,
+} from "./agent-session-queue";
 import type {
 	ActiveAdvisor,
 	AdvisorRuntimeDescriptor,
@@ -363,7 +373,6 @@ import {
 	type InterruptedThinkingDetails,
 	isUserInterruptAbort,
 	type PythonExecutionMessage,
-	readQueueChipText,
 	SILENT_ABORT_MARKER,
 	SKILL_PROMPT_MESSAGE_TYPE,
 	stripImagesFromMessage,
@@ -381,6 +390,8 @@ import { ToolChoiceQueue } from "./tool-choice-queue";
 import { planTurnPersistence, sameMessageContent, sessionMessagePersistenceKey } from "./turn-persistence";
 import { classifyUnexpectedStop, isUnexpectedStopCandidate } from "./unexpected-stop-classifier";
 import { YieldQueue } from "./yield-queue";
+
+export type { RestoredQueuedMessage } from "./agent-session-queue";
 
 const SESSION_STOP_CONTINUATION_CAP = 8;
 const PLAN_MODE_REMINDER_MAX = 3;
@@ -854,88 +865,6 @@ function createHandoffFileName(date = new Date()): string {
 // ============================================================================
 // AgentSession Class
 // ============================================================================
-
-/** Entry returned by {@link AgentSession.clearQueue} / {@link AgentSession.popLastQueuedMessage}. */
-export type RestoredQueuedMessage = { text: string; images?: ImageContent[] };
-
-function queuedTextContent(message: AgentMessage): string | undefined {
-	if (!("content" in message)) return undefined;
-	const content = message.content;
-	if (typeof content === "string") return content;
-	return content.find((part): part is TextContent => part.type === "text")?.text;
-}
-
-function queuedImageContent(message: AgentMessage): ImageContent[] | undefined {
-	if (!("content" in message) || typeof message.content === "string") return undefined;
-	const images = message.content.filter(
-		(part): part is ImageContent =>
-			part.type === "image" && typeof part.data === "string" && typeof part.mimeType === "string",
-	);
-	return images.length > 0 ? images : undefined;
-}
-
-function isDisplayableQueuedMessage(message: AgentMessage): boolean {
-	return !(message.role === "custom" && message.display === false);
-}
-
-function isAdvisorCard(message: AgentMessage): message is CustomMessage {
-	return message.role === "custom" && message.customType === "advisor";
-}
-
-/**
- * A queued message the user can restore to the editor / pull back as a draft.
- * Only genuinely user-authored messages qualify: plain user turns, or custom
- * messages explicitly attributed to the user (e.g. `/skill` invocations).
- * Agent-authored queued cards — advisor concern/blocker notes, IRC asides,
- * extension notices, hidden goal/plan/budget steers — ride the same
- * steer/follow-up queues but must never be dumped into the editor on Esc/Alt+Up.
- */
-function isUserQueuedMessage(message: AgentMessage): boolean {
-	if (message.role === "user") return true;
-	return message.role === "custom" && message.attribution === "user" && message.display !== false;
-}
-
-/** Custom-message types of the hidden magic-keyword notices that `#createMagicKeywordNotices`
- *  enqueues alongside a user prompt. Keep in sync with that method. */
-const MAGIC_KEYWORD_NOTICE_TYPES: ReadonlySet<string> = new Set([
-	"ultrathink-notice",
-	"orchestrate-notice",
-	"workflow-notice",
-]);
-
-/** Custom-message type of the hidden companion carrying vision descriptions of image
- *  attachments sent to a text-only model (see `#buildImageDescriptionNotice`). */
-const IMAGE_ATTACHMENT_DESCRIPTION_TYPE = "image-attachment-description";
-
-/**
- * A hidden, user-attributed companion of a queued user prompt: the magic-keyword
- * notices (`ultrathink`/`orchestrate`/`workflow`) enqueued alongside the user
- * message. They are `attribution: "user"` but `display: false`, so they are not
- * editor-restorable; when the user pulls their prompt back out of the queue these
- * must leave with it rather than linger as stale, companion-less steering. Scoped to
- * the known notice types so an unrelated hidden user custom is never silently dropped.
- */
-function isHiddenUserCompanion(message: AgentMessage): boolean {
-	return (
-		message.role === "custom" &&
-		message.attribution === "user" &&
-		message.display === false &&
-		(MAGIC_KEYWORD_NOTICE_TYPES.has(message.customType) || message.customType === IMAGE_ATTACHMENT_DESCRIPTION_TYPE)
-	);
-}
-
-function queueChipText(message: AgentMessage): string {
-	if (message.role === "custom") {
-		return readQueueChipText(message.details) ?? queuedTextContent(message) ?? "";
-	}
-	const text = queuedTextContent(message) ?? "";
-	if (text) return text;
-	return queuedImageContent(message) ? "[Image]" : "";
-}
-
-function toRestoredQueuedMessage(message: AgentMessage): RestoredQueuedMessage {
-	return { text: queueChipText(message), images: queuedImageContent(message) };
-}
 
 function mergeLlmCompactionPreserveData(
 	hookPreserveData: Record<string, unknown> | undefined,
